Move the TestCx struct to its own file.
